+++ lib/provider/filter_provider.dart
@@ -98,10 +98,10 @@ class FilterProvider extends ChangeNotifier implements EventFilter {
       return true;
     }
 
-    if (EventKind.SUPPORTED_EVENTS.contains(e.kind) ||
-        e.kind == EventKind.DIRECT_MESSAGE ||
-        e.kind == EventKind.GIFT_WRAP) {
-      if (e.kind != EventKind.ZAP_GOALS && !wotProvider.check(e.pubkey)) {
+    if (EventKind.supportedEvents.contains(e.kind) ||
+        e.kind == EventKind.directMessage ||
+        e.kind == EventKind.giftWrap) {
+      if (e.kind != EventKind.zapGoals && !wotProvider.check(e.pubkey)) {
         return true;
       }